+++ src/ser/mod.rs
@@ -17,7 +17,7 @@ where
     W: io::Write,
     T: ?Sized + Serialize,
 {
-    let mut s = Serializer::new(writer, None, false)?;
+    let mut s = Serializer::new(writer, None)?;
     value.serialize(&mut s)
 }
 
@@ -27,7 +27,7 @@ where
     W: io::Write,
     T: ?Sized + Serialize,
 {
-    let mut s = Serializer::new(writer, Some(config), false)?;
+    let mut s = Serializer::new(writer, Some(config))?;
     value.serialize(&mut s)
 }
 
@@ -40,7 +40,7 @@ where
     T: ?Sized + Serialize,
 {
     let buf = Vec::new();
-    let mut s = Serializer::new(buf, None, false)?;
+    let mut s = Serializer::new(buf, None)?;
     value.serialize(&mut s)?;
     Ok(String::from_utf8(s.output).expect("Ron should be utf-8"))
 }
@@ -51,7 +51,7 @@ where
     T: ?Sized + Serialize,
 {
     let buf = Vec::new();
-    let mut s = Serializer::new(buf, Some(config), false)?;
+    let mut s = Serializer::new(buf, Some(config))?;
     value.serialize(&mut s)?;
     Ok(String::from_utf8(s.output).expect("Ron should be utf-8"))
 }
@@ -86,8 +86,10 @@ pub struct PrettyConfig {
     #[serde(default = "default_indentor")]
     pub indentor: String,
     /// Separator string
+    #[serde(default = "default_separator")]
     pub separator: String,
     // Whether to emit struct names
+    #[serde(default = "default_struct_names")]
     pub struct_names: bool,
     /// Separate tuple members with indentation
     #[serde(default = "default_separate_tuple_members")]
@@ -243,6 +245,14 @@ fn default_indentor() -> String {
     "    ".to_string()
 }
 
+fn default_separator() -> String {
+    "    ".to_string()
+}
+
+fn default_struct_names() -> bool {
+    false
+}
+
 fn default_separate_tuple_members() -> bool {
     false
 }
@@ -257,12 +267,12 @@ impl Default for PrettyConfig {
             depth_limit: default_depth_limit(),
             new_line: default_new_line(),
             indentor: default_indentor(),
+            struct_names: default_struct_names(),
             separate_tuple_members: default_separate_tuple_members(),
             enumerate_arrays: default_enumerate_arrays(),
             extensions: Extensions::default(),
             decimal_floats: default_decimal_floats(),
             separator: String::from(" "),
-            struct_names: false,
             compact_arrays: false,
             _future_proof: (),
         }
@@ -276,7 +286,6 @@ impl Default for PrettyConfig {
 pub struct Serializer<W: io::Write> {
     output: W,
     pretty: Option<(PrettyConfig, Pretty)>,
-    struct_names: bool,
     is_empty: Option<bool>,
 }
 
@@ -284,7 +293,7 @@ impl<W: io::Write> Serializer<W> {
     /// Creates a new `Serializer`.
     ///
     /// Most of the time you can just use `to_string` or `to_string_pretty`.
-    pub fn new(mut writer: W, config: Option<PrettyConfig>, struct_names: bool) -> Result<Self> {
+    pub fn new(mut writer: W, config: Option<PrettyConfig>) -> Result<Self> {
         if let Some(conf) = &config {
             if conf.extensions.contains(Extensions::IMPLICIT_SOME) {
                 writer.write_all(b"#![enable(implicit_some)]")?;
@@ -306,7 +315,6 @@ impl<W: io::Write> Serializer<W> {
                     },
                 )
             }),
-            struct_names,
             is_empty: None,
         })
     }
@@ -396,6 +404,13 @@ impl<W: io::Write> Serializer<W> {
         }
         self.output.write_all(name.as_bytes())?;
         Ok(())
+    }
+
+    fn struct_names(&self) -> bool {
+        self.pretty
+            .as_ref()
+            .map(|(pc, _)| pc.struct_names)
+            .unwrap_or(false)
     }
 }
 
@@ -529,7 +544,7 @@ impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
     }
 
     fn serialize_unit_struct(self, name: &'static str) -> Result<()> {
-        if self.struct_names {
+        if self.struct_names() {
             self.write_identifier(name)?;
 
             Ok(())
@@ -552,7 +567,7 @@ impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
             return value.serialize(&mut *self);
         }
 
-        if self.struct_names {
+        if self.struct_names() {
             self.write_identifier(name)?;
         }
 
@@ -622,7 +637,7 @@ impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
         name: &'static str,
         len: usize,
     ) -> Result<Self::SerializeTupleStruct> {
-        if self.struct_names {
+        if self.struct_names() {
             self.write_identifier(name)?;
         }
 
@@ -667,7 +682,7 @@ impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
     }
 
     fn serialize_struct(self, name: &'static str, len: usize) -> Result<Self::SerializeStruct> {
-        if self.struct_names {
+        if self.struct_names() {
             self.write_identifier(name)?;
         }
         self.output.write_all(b"(")?;